backend/requirements.txt
Django==5.0.0
graphene-django==3.2.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
gunicorn==21.2.0
graphql-core==3.2.3